rational: use true min/max values for extended range
Will address issues caused by comparisons to our MIN/MAX constants when
a numerator/denominator is > INT32_MAX/MIN.

+++ app/common/rational.h
@@ -127,9 +127,8 @@ private:
   static intType gcd(const intType &x, const intType &y);
 };
 
-// We define these limits at 32-bit to try avoiding integer overflow
-#define RATIONAL_MIN rational(INT32_MIN, 1)
-#define RATIONAL_MAX rational(INT32_MAX, 1)
+#define RATIONAL_MIN rational(INT64_MIN, 1)
+#define RATIONAL_MAX rational(INT64_MAX, 1)
 
 uint qHash(const rational& r, uint seed);
